+++ modules/users/create_user.sh
@@ -2,5 +2,9 @@ create_user () {
     user=$1
     password=$2
 
-    sudo useradd $user
+    sudo useradd --shell /bin/bash $user
+    sudo mkdir /home/$user
+
+    sudo chown -R $user:$user /home/$user
+    sudo mkdir -p /home/$user/.ssh
 }