+++ CHANGELOG.md
@@ -9,9 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Added
+- [2024-04-28] Add support for .bashrc if it doesnt exist
 - [2024-04-28] Add admin user creation
 - [2024-04-28] Add pretty print to the installations and init_fs processes
 - [2024-04-28] Add filesystem initialization for vm folders
 - [2024-04-28] Add build script that bundles the file into build/index.sh
+
+
+### Changed
+- [2024-04-28] Changed init_fs from coretools to admin home
 
 ## [0.1.0] - 2024-04-28 (The Initial Update)